+++ test/cases/phase1/const/infer/expected.c
@@ -2,7 +2,7 @@
 
 int main(void) {
     TSC_INIT();
-    const int32_t x = 100;
-    printf("%d\n", x);
+    const double x = 100.0;
+    printf("%g\n", x);
     return 0;
 }
