+++ test/cases/phase1/logical/and-returns-operand/expected.c
@@ -2,7 +2,7 @@
 
 int main(void) {
     TSC_INIT();
-    int32_t _tsc_lhs_0 = 1;
+    double _tsc_lhs_0 = 1;
     int32_t x = (_tsc_lhs_0) ? 42 : _tsc_lhs_0;
     printf("%d\n", x);
     return 0;
