+++ test/cases/phase2/generics/bound/expected.c
@@ -1,6 +1,6 @@
 #include "runtime.h"
 
-typedef struct { int32_t id; String name; } _anon_0;
+typedef struct { double id; String name; } _anon_0;
 
 int32_t getId__anon_0(_anon_0 obj) {
     return obj.id;
