+++ test/cases/phase2/type-inference/from-return/expected.c
@@ -1,12 +1,12 @@
 #include "runtime.h"
 
-int32_t getValue(void) {
+double getValue(void) {
     return 100;
 }
 
 int main(void) {
     TSC_INIT();
-    const int32_t x = getValue();
-    printf("%d\n", x);
+    const double x = getValue();
+    printf("%g\n", x);
     return 0;
 }
