+++ test/cases/phase3/destructuring/with-defaults/expected.c
@@ -2,11 +2,11 @@
 
 int main(void) {
     TSC_INIT();
-    const int32_t _obj_x = 10;
-    const int32_t _obj_y = 0;
-    const int32_t x = _obj_x;
-    const int32_t y = (_obj_y != 0) ? _obj_y : 5;
-    printf("%d\n", x);
-    printf("%d\n", y);
+    const double _obj_x = 10;
+    const double _obj_y = 0;
+    const double x = _obj_x;
+    const double y = (_obj_y != 0) ? _obj_y : 5;
+    printf("%g\n", x);
+    printf("%g\n", y);
     return 0;
 }
