+++ test/cases/phase4/closures/closure-pass-to-fn/expected.c
@@ -1,6 +1,6 @@
 #include "runtime.h"
 
-typedef struct { int32_t x; int32_t y; } _closure_0_env;
+typedef struct { double x; double y; } _closure_0_env;
 
 static int32_t _closure_0_fn(_closure_0_env *env) {
     return env->x + env->y;
@@ -8,8 +8,8 @@ static int32_t _closure_0_fn(_closure_0_env *env) {
 
 int main(void) {
     TSC_INIT();
-    int32_t x = 7;
-    int32_t y = 3;
+    double x = 7.0;
+    double y = 3.0;
     _closure_0_env add_env = {.x = x, .y = y};
     tsc_closure add = {.env = &add_env, .fn = (void*)_closure_0_fn};
     printf("%d\n", ((int32_t (*)(void *))add.fn)(add.env));
